7.4 update

+++ StatusNode/StatusNode.cs
@@ -171,7 +171,7 @@ namespace EnemyListDebuffs.StatusNode
             newImageNode->AtkResNode.DrawFlags = 0;
 
             newImageNode->WrapMode = 1;
-            newImageNode->Flags |= (byte)ImageNodeFlags.AutoFit;
+            newImageNode->Flags |= ImageNodeFlags.AutoFit;
 
             var partsList = (AtkUldPartsList*)IMemorySpace.GetUISpace()->Malloc((ulong)sizeof(AtkUldPartsList), 8);
             if (partsList == null)
